consolidate wasm/native setup
