+++ widgetry/src/backend_glow_wasm.rs
@@ -1,11 +1,10 @@
-use glow::HasContext;
 use std::rc::Rc;
 use wasm_bindgen::JsCast;
 use winit::platform::web::WindowExtWebSys;
 
 use abstutil::Timer;
 
-use crate::backend_glow::{GfxCtxInnards, PrerenderInnards, SpriteTexture};
+use crate::backend_glow::{build_program, GfxCtxInnards, PrerenderInnards, SpriteTexture};
 use crate::ScreenDims;
 
 pub fn setup(
@@ -88,17 +87,13 @@ pub fn setup(
             .dyn_into::<web_sys::WebGl2RenderingContext>()
             .map_err(|err| anyhow!("unable to cast to WebGl2RenderingContext. error: {:?}", err))?;
         let gl = glow::Context::from_webgl2_context(webgl2_context);
-        let shader_inputs = [
-            (
-                glow::VERTEX_SHADER,
+        let program = unsafe {
+            build_program(
+                &gl,
                 include_str!("../shaders/vertex_300.glsl"),
-            ),
-            (
-                glow::FRAGMENT_SHADER,
                 include_str!("../shaders/fragment_300.glsl"),
-            ),
-        ];
-        let program = unsafe { build_program(&gl, &shader_inputs)? };
+            )?
+        };
 
         timer.start("load textures");
         let sprite_texture = SpriteTexture::new(
@@ -128,18 +123,13 @@ pub fn setup(
             .dyn_into::<web_sys::WebGlRenderingContext>()
             .map_err(|err| anyhow!("unable to cast to WebGlRenderingContext. error: {:?}", err))?;
         let gl = glow::Context::from_webgl1_context(webgl1_context);
-
-        let shader_inputs = [
-            (
-                glow::VERTEX_SHADER,
+        let program = unsafe {
+            build_program(
+                &gl,
                 include_str!("../shaders/vertex_webgl1.glsl"),
-            ),
-            (
-                glow::FRAGMENT_SHADER,
                 include_str!("../shaders/fragment_webgl1.glsl"),
-            ),
-        ];
-        let program = unsafe { build_program(&gl, &shader_inputs)? };
+            )?
+        };
 
         timer.start("load textures");
         let sprite_texture = SpriteTexture::new(
@@ -154,58 +144,6 @@ pub fn setup(
         timer.stop("load textures");
 
         Ok((program, gl))
-    }
-
-    // TODO: move this to backend_glow and share w/ native backend?
-    /// shaders_input: (shader_type: u32, shader_src: &str)
-    unsafe fn build_program(
-        gl: &glow::Context,
-        shaders_input: &[(u32, &str)],
-    ) -> anyhow::Result<glow::Program> {
-        let program = gl.create_program().expect("Cannot create program");
-
-        let shaders = shaders_input
-            .iter()
-            .map(|(shader_type, source)| {
-                let shader = gl
-                    .create_shader(*shader_type)
-                    .expect("Cannot create shader");
-                gl.shader_source(shader, source);
-                gl.compile_shader(shader);
-                if !gl.get_shader_compile_status(shader) {
-                    error!("Shader error: {}", gl.get_shader_info_log(shader));
-                    panic!(gl.get_shader_info_log(shader));
-                }
-                gl.attach_shader(program, shader);
-                shader
-            })
-            .collect::<Vec<_>>();
-
-        gl.link_program(program);
-        if !gl.get_program_link_status(program) {
-            error!("Linking error: {}", gl.get_program_info_log(program));
-            panic!(gl.get_program_info_log(program));
-        }
-        for shader in shaders {
-            gl.detach_shader(program, shader);
-            gl.delete_shader(shader);
-        }
-        gl.use_program(Some(program));
-
-        gl.enable(glow::SCISSOR_TEST);
-
-        gl.enable(glow::DEPTH_TEST);
-        gl.depth_func(glow::LEQUAL);
-
-        gl.enable(glow::BLEND);
-        gl.blend_func_separate(
-            glow::ONE,
-            glow::ONE_MINUS_SRC_ALPHA,
-            glow::ONE_MINUS_DST_ALPHA,
-            glow::ONE,
-        );
-
-        Ok(program)
     }
 
     (
